requirements.txt
numpy==1.24.4
accelerate==0.30.1
imageio==2.37.0
av==14.4.0
aiofiles==23.2.1
transformers==4.47.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
huggingface_hub==0.24.6
opencv-python
matplotlib
albucore==0.0.24
albumentations==2.0.6
einops==0.8.1
open_clip_torch
kornia
lpips
httpx
click
ninja==1.11.1.3
jinja2==3.1.4 
pillow==9.4.0
xformers==0.0.20
yaml==0.2.5
basicsr
gradio==4.44.1
scikit-image==0.24.0
transformers==4.48.0
triton==2.0.0